Рефакторинг: разделение модуля util.js
